requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
coverage==7.6.12
decorator==5.2.1
executing==2.2.0
filelock==3.17.0
Flask==3.1.0
fsspec==2025.2.0
gemini_ai==0.0.5
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.3.0
google-generativeai==0.8.4
googleapis-common-protos==1.68.0
grpcio==1.71.0rc2
grpcio-status==1.71.0rc2
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
iniconfig==2.0.0
ipython==8.32.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jupyterlab_widgets==3.0.13
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
packaging==24.2
parso==0.8.4
pillow==11.1.0
pluggy==1.5.0
prompt_toolkit==3.0.50
proto-plus==1.26.0
protobuf==5.29.3
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
pytest==8.3.4
pytest-cov==6.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.5.2
setuptools==75.8.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.13.1
tokenizers==0.21.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
websockets==14.2
Werkzeug==3.1.3
widgetsnbextension==4.0.13
